StationReader and Task
Added new variable speed to Task. a new Constructor with speed added to Task. StationReader is now constructing Tasks with speed if there is no speed given speed is set to 1 wich is default. An error message will be printed out if that is the case. StationReader no longer checks Tasks for their size.