Station and subStations
The methods of these claseses are updated. CalculateStartTime method will be used sellecting optimal station for a task.First EventFlow add the task to station targetTasks list with addTask method.Then if the event of the task end, the next Task will pick from the station by pickTask method. The duration of the task is calculating by the formula given at the description.